requirements.txt


python-box
requests
tqdm
pandas
numpy
